Preparing the first snapshot
=== The changes ===
Improvements in the system of logs and implementation of rotation of logs.
Improvements in the orms and in the api.
Creation of the express route manager.
Creation of new utils such as Math and TaskTimer.
Creation of a configuration to render static routes in the chat.
Creation of authentication system.
Improvements in the sql initiation script.